scrapeData.py
import time
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.keys import Keys
from bs4 import BeautifulSoup
import os
import re

# ... (rest of your scraping logic here)

# Example placeholder for the main scraping function
def scrape_linkedin_posts(profile_url, session_cookie, num_posts=10):
    """
    Scrape the latest LinkedIn posts for a user using Selenium and session cookie.
    """
    # Your scraping logic here
    pass

if __name__ == "__main__":
    # Example usage
    profile_url = "https://www.linkedin.com/in/your-profile/"
    session_cookie = "your_session_cookie_here"
    posts = scrape_linkedin_posts(profile_url, session_cookie)
    print(posts) 