requirements-lightweight.txt
flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
tavily-python==0.2.8
openai==1.3.0
waitress==3.0.2
Werkzeug==2.3.7
numpy==1.24.3
python-dotenv==1.0.0 